Fixes #115 , Fixes #50
Ladder grabbing outside the screen is like in the original games now
(only grab at the border and starts the transition).
Middle walk sprite is used when landing like in the original games
(almost imperceptible).
#104 (moving platforms) seems more rare now, albeit not entirely fixed.
Hint: GM features are there to be used. If something is solid, check the
Solid black in the object. Let GM handle absurd collision cases.
Hint 2: motion planning (mp_*) functions shouldn't be used for enemies
that fly through walls. It is used for planning around obstacles (had to
fix battons).